sbnb-configure-networking.sh: configure the host networking for the VM
This script is used to configure the host networking for the VM.
It creates a bridge interface (br0) and attaches the main interface to it.
This allows the VM to be attached to the host and obtain an IP address from the same DHCP server as the host itself.
This also allows the VM to be directly accessible from the host and other devices on the LAN.
The script is idempotent and can be run multiple times without causing issues.

Warning: This script will overwrite the existing network configuration on the host (rm -f /etc/systemd/network/*)

Below is an example of the network configuration after running this script and starting two VMs:

+---------------------+
|        LAN          |
+---------------------+
          |
          |
+---------------------+
|        Host         |
| +-----------------+ |
| |     br0         | |
| +-----------------+ |
|   |           |     |
| +-----+     +-----+ |
| | tap0|     | tap1| |
| +-----+     +-----+ |
|   |           |     |
| +-----+     +-----+ |
| | VM1 |     | VM2 | |
| +-----+     +-----+ |
+---------------------+