sbnb-configure-storage.sh: configure the host storage
This script is used to configure the host storage for the VM.  It finds
all drives that don't have any partitions on them, allocates all
100%FREE space to LVM, creates one flat space from all available drives
(similar to RAID0), creates PV, VG, LV, formats LV to ext4 with "-m 0"
and mounts it to /mnt/sbnb-data. The script is idempotent and can be
run multiple times without causing issues.

Warning: This script will overwrite the existing storage configuration on the host.

Example representation of the final state with two 1TB nvme drives

Initial state:
/dev/nvme0n1 (1TB)
/dev/nvme1n1 (1TB)

After running the script:
/dev/nvme0n1 (1TB)  /dev/nvme1n1 (1TB)
     |                   |
     v                   v
  pvcreate            pvcreate
     |                   |
     v                   v
  sbnb-vg (Volume Group, 2TB)
     |
     v
  sbnb-lv (Logical Volume, 2TB)
     |
     v
  ext4 filesystem
     |
     v
  /mnt/sbnb-data (Mount Point, 2TB)